refactor: [CI-CNI] Removing the predefined interfaces from validate package
