+++ test/validate/windows_validate.go
@@ -1,20 +1,10 @@
 package validate
 
 import (
-	"context"
 	"encoding/json"
-	"log"
 	"net"
 
-	k8sutils "github.com/Azure/azure-container-networking/test/internal/k8sutils"
 	"github.com/pkg/errors"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-)
-
-const (
-	privilegedWindowsDaemonSetPath = "../manifests/load/privileged-daemonset-windows.yaml"
-	windowsNodeSelector            = "kubernetes.io/os=windows"
 )
 
 var (
@@ -23,12 +13,15 @@ var (
 	azureVnetIpamCmd = []string{"powershell", "-c", "cat ../../k/azure-vnet-ipam.json"}
 )
 
-type WindowsValidator struct {
-	clientset   *kubernetes.Clientset
-	config      *rest.Config
-	namespace   string
-	cni         string
-	restartCase bool
+var windowsChecksMap = map[string][]check{
+	"cniv1": {
+		{"hns", hnsStateFileIps, privilegedLabelSelector, privilegedNamespace, hnsEndPointCmd},
+		{"azure-vnet", azureVnetIps, privilegedLabelSelector, privilegedNamespace, azureVnetCmd},
+		{"azure-vnet-ipam", azureVnetIpamIps, privilegedLabelSelector, privilegedNamespace, azureVnetIpamCmd},
+	},
+	"cniv2": {
+		{"azure-vnet", azureVnetIps, privilegedLabelSelector, privilegedNamespace, azureVnetCmd},
+	},
 }
 
 type HNSEndpoint struct {
@@ -78,60 +71,6 @@ type AddressPool struct {
 type AddressRecord struct {
 	Addr  net.IP
 	InUse bool
-}
-
-type check struct {
-	name             string
-	stateFileIps     func([]byte) (map[string]string, error)
-	podLabelSelector string
-	podNamespace     string
-	cmd              []string
-}
-
-func CreateWindowsValidator(ctx context.Context, clienset *kubernetes.Clientset, config *rest.Config, namespace, cni string, restartCase bool) (*WindowsValidator, error) {
-	// deploy privileged pod
-	privilegedDaemonSet, err := k8sutils.MustParseDaemonSet(privilegedWindowsDaemonSetPath)
-	if err != nil {
-		return nil, errors.Wrap(err, "unable to parse daemonset")
-	}
-	daemonsetClient := clienset.AppsV1().DaemonSets(privilegedNamespace)
-	if err := k8sutils.MustCreateDaemonset(ctx, daemonsetClient, privilegedDaemonSet); err != nil {
-		return nil, errors.Wrap(err, "unable to create daemonset")
-	}
-	if err := k8sutils.WaitForPodsRunning(ctx, clienset, privilegedNamespace, privilegedLabelSelector); err != nil {
-		return nil, errors.Wrap(err, "error while waiting for pods to be running")
-	}
-	return &WindowsValidator{
-		clientset:   clienset,
-		config:      config,
-		namespace:   namespace,
-		cni:         cni,
-		restartCase: restartCase,
-	}, nil
-}
-
-func (v *WindowsValidator) ValidateStateFile(ctx context.Context) error {
-	checkSet := make(map[string][]check) // key is cni type, value is a list of check
-
-	checkSet["cniv1"] = []check{
-		{"hns", hnsStateFileIps, privilegedLabelSelector, privilegedNamespace, hnsEndPointCmd},
-		{"azure-vnet", azureVnetIps, privilegedLabelSelector, privilegedNamespace, azureVnetCmd},
-		{"azure-vnet-ipam", azureVnetIpamIps, privilegedLabelSelector, privilegedNamespace, azureVnetIpamCmd},
-	}
-
-	checkSet["cniv2"] = []check{
-		{"azure-vnet", azureVnetIps, privilegedLabelSelector, privilegedNamespace, azureVnetCmd},
-	}
-
-	// this is checking all IPs of the pods with the statefile
-	for _, check := range checkSet[v.cni] {
-		err := v.validateIPs(ctx, check.stateFileIps, check.cmd, check.name, check.podNamespace, check.podLabelSelector)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 func hnsStateFileIps(result []byte) (map[string]string, error) {
@@ -189,47 +128,4 @@ func azureVnetIpamIps(result []byte) (map[string]string, error) {
 		}
 	}
 	return azureVnetIpamPodIps, nil
-}
-
-func (v *WindowsValidator) validateIPs(ctx context.Context, stateFileIps stateFileIpsFunc, cmd []string, checkType, namespace, labelSelector string) error {
-	log.Println("Validating ", checkType, " state file")
-	nodes, err := k8sutils.GetNodeListByLabelSelector(ctx, v.clientset, windowsNodeSelector)
-	if err != nil {
-		return errors.Wrapf(err, "failed to get node list")
-	}
-	for index := range nodes.Items {
-		// get the privileged pod
-		pod, err := k8sutils.GetPodsByNode(ctx, v.clientset, namespace, labelSelector, nodes.Items[index].Name)
-		if err != nil {
-			return errors.Wrapf(err, "failed to get privileged pod")
-		}
-		podName := pod.Items[0].Name
-		// exec into the pod to get the state file
-		result, err := k8sutils.ExecCmdOnPod(ctx, v.clientset, namespace, podName, cmd, v.config)
-		if err != nil {
-			return errors.Wrapf(err, "failed to exec into privileged pod")
-		}
-		filePodIps, err := stateFileIps(result)
-		if err != nil {
-			return errors.Wrapf(err, "failed to get pod ips from state file")
-		}
-		if len(filePodIps) == 0 && v.restartCase {
-			log.Printf("No pods found on node %s", nodes.Items[index].Name)
-			continue
-		}
-		// get the pod ips
-		podIps := getPodIPsWithoutNodeIP(ctx, v.clientset, nodes.Items[index])
-
-		check := compareIPs(filePodIps, podIps)
-
-		if !check {
-			return errors.Wrapf(errors.New("State file validation failed"), "for %s on node %s", checkType, nodes.Items[index].Name)
-		}
-	}
-	log.Printf("State file validation for %s passed", checkType)
-	return nil
-}
-
-func (v *WindowsValidator) ValidateRestartNetwork(context.Context) error {
-	return errors.New("not implemented")
 }
